% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_classes.R
\name{determine_classes}
\alias{determine_classes}
\title{Determine unique classes (internal function)}
\usage{
determine_classes(x, window, window_size)
}
\arguments{
\item{x}{\itemize{
\item a \code{stars} or \code{stars_proxy} object
}}

\item{window}{\itemize{
\item a \code{sf} object
}}

\item{window_size}{\itemize{
\item a single value numeric object
}}
}
\value{
A list with vector of numbers (unique classes)
}
\description{
Determine unique classes (internal function)
}
