% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cocoma.R
\name{get_cocoma}
\alias{get_cocoma}
\alias{get_cocoma.RasterLayer}
\title{Create a co-located co-occurrence matrix (cocoma)}
\usage{
get_cocoma(x, y, neighbourhood, size, shift)

\method{get_cocoma}{RasterLayer}(x, y, neighbourhood = 4, size = NULL,
  shift = NULL)
}
\arguments{
\item{x}{A RasterLayer with categories}

\item{y}{A RasterLayer with categories}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}
}
\value{
A co-located co-occurrence matrix
}
\description{
Create a co-located co-occurrence matrix (cocoma)
}
\examples{
library(comat)
library(raster)
l1 = raster(matrix(sample(1:2, size = 10000, replace = TRUE), ncol = 100))
l2 = raster(matrix(sample(c(9, 6, 3), size = 10000, replace = TRUE), ncol = 100))

coom = get_cocoma(l1, l2)
coom

coom2 = get_cocoma(l1, l2, size = 10, shift = 10)
coom2
}
