% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incoma.R
\name{get_incoma}
\alias{get_incoma}
\alias{get_incoma.RasterStack}
\title{Create an integrated co-occurrence matrix (wecoma)}
\usage{
get_incoma(x, neighbourhood, size, shift)

\method{get_incoma}{RasterStack}(x, neighbourhood = 4, size = NULL,
  shift = NULL)
}
\arguments{
\item{x}{A RasterStack object containing categorical rasters}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}

\item{size}{Expressed in the numbers of cells, is a length of the side of a square-shaped block of cells. It defines the extent of a local pattern. If \code{size=NULL} calculations are performed for a whole area}

\item{shift}{Defines the shift between adjacent squares of cells along with the N-S and W-E directions. It describes the density (resolution) of the output grid. The resolution of the output map will be reduced to the original resolution multiplied by the shift. If shift=size the input map will be divided into a grid of non-overlapping square windows. Each square window defines the extent of a local pattern. If shift < size - results in the grid of overlapping square windows.}
}
\value{
An integrated co-occurrence matrix
}
\description{
Create an integrated co-occurrence matrix (wecoma)
}
\examples{
library(raster)

set.seed(10100)
l1 = raster(matrix(sample(1:2, size = 1000000, replace = TRUE), ncol = 1000))
l2 = raster(matrix(sample(c(9, 6, 3), size = 1000000, replace = TRUE), ncol = 1000))
x = stack(l1, l2, l1)

get_incoma(x)
get_incoma(x, size = 100, shift = 100)
}
