% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incove.R
\name{get_incove}
\alias{get_incove}
\title{Create an integrated co-occurrence vector (wecove)}
\usage{
get_incove(x, ordered = TRUE)
}
\arguments{
\item{x}{A matrix - an output of the \code{\link[=get_incoma]{get_incoma()}} function}

\item{ordered}{The type of pairs considered.
Either ordered (TRUE) or unordered (FALSE).
The default is TRUE.}
}
\value{
An integrated co-occurrence vector
}
\description{
Converts an integrated co-occurrence matrix (wecoma) to
an integrated co-occurrence vector (wecove)
}
\examples{
library(comat)
library(raster)

l1 = raster(matrix(sample(1:2, size = 100, replace = TRUE), ncol = 10))
l2 = raster(matrix(sample(c(9, 6, 3), size = 100, replace = TRUE), ncol = 10))
x = stack(l1, l2, l1)

incom = get_incoma(x)

incov1 = get_incove(incom)
incov1

incov2 = get_incove(incom, ordered = FALSE)
incov2
}
