% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wecoma.R
\name{get_wecoma}
\alias{get_wecoma}
\alias{get_wecoma.RasterLayer}
\title{Create a weighted co-occurrence matrix (wecoma)}
\usage{
get_wecoma(x, w, neighbourhood, size, shift, fun, na_action)

\method{get_wecoma}{RasterLayer}(x, w, neighbourhood = 4, size = NULL,
  shift = NULL, fun = "mean", na_action = "replace")
}
\arguments{
\item{x}{A RasterLayer with categories}

\item{w}{A RasterLayer with weights}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}

\item{size}{Expressed in the numbers of cells, is a length of the side of a square-shaped block of cells. It defines the extent of a local pattern. If \code{size=NULL} calculations are performed for a whole area}

\item{shift}{Defines the shift between adjacent squares of cells along with the N-S and W-E directions. It describes the density (resolution) of the output grid. The resolution of the output map will be reduced to the original resolution multiplied by the shift. If shift=size the input map will be divided into a grid of non-overlapping square windows. Each square window defines the extent of a local pattern. If shift < size - results in the grid of overlapping square windows.}

\item{fun}{"mean", "geometric_mean", or "focal". The default is "mean".}

\item{na_action}{"replace", "omit", or "keep". The default is "replace".}
}
\value{
A weighted co-occurrence matrix
}
\description{
Create a weighted co-occurrence matrix (wecoma)
}
\examples{
library(comat)
library(raster)
data(x, package = "comat")
data(w, package = "comat")
plot(x)
plot(w)

wom = get_wecoma(x, w)
wom
wom2 = get_wecoma(x, w, size = 2, shift = 2)
wom2
}
