% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_cocoma.R
\name{lop_cocoma}
\alias{lop_cocoma}
\alias{lop_cocoma.RasterLayer}
\title{Create a co-located co-occurrence matrix (cocoma)}
\usage{
lop_cocoma(x, y, neighbourhood, size, shift, threshold)

\method{lop_cocoma}{RasterLayer}(x, y, neighbourhood = 4, size = NULL, shift = NULL, threshold = 0.5)
}
\arguments{
\item{x}{A RasterLayer with categories}

\item{y}{A RasterLayer with categories}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}

\item{size}{Expressed in the numbers of cells, is a length of the side of a square-shaped block of cells. It defines the extent of a local pattern. If \code{size=NULL} calculations are performed for a whole area}

\item{shift}{Defines the shift between adjacent squares of cells along with the N-S and W-E directions. It describes the density (resolution) of the output grid. The resolution of the output map will be reduced to the original resolution multiplied by the shift. If shift=size the input map will be divided into a grid of non-overlapping square windows. Each square window defines the extent of a local pattern. If shift < size - results in the grid of overlapping square windows.}

\item{threshold}{The share of NA cells to allow calculation in a square-shaped window.}
}
\value{
A co-located co-occurrence matrix
}
\description{
Create a co-located co-occurrence matrix (cocoma)
}
\examples{
library(comat)
library(raster)
l1 = raster(matrix(sample(1:2, size = 10000, replace = TRUE), ncol = 100))
l2 = raster(matrix(sample(c(9, 6, 3), size = 10000, replace = TRUE), ncol = 100))

coom = lop_cocoma(l1, l2)
coom

coom2 = lop_cocoma(l1, l2, size = 10, shift = 10)
coom2
}
