% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_cocove.R
\name{lop_cocove}
\alias{lop_cocove}
\alias{lop_cocove.cocoma}
\title{Create a co-located co-occurrence vector (cocove)}
\usage{
lop_cocove(x, ordered, normalization)

\method{lop_cocove}{cocoma}(x, ordered = TRUE, normalization = "none")
}
\arguments{
\item{x}{A matrix - an output of the \code{\link[=lop_cocoma]{lop_cocoma()}} function}

\item{ordered}{The type of pairs considered.
Either "ordered" (TRUE) or "unordered" (FALSE).
The default is TRUE.}

\item{normalization}{Should the output vector be normalized?
Either "none" or "pdf".
The "pdf" option normalizes a vector to sum to one.
The default is "none".}
}
\value{
A co-located co-occurrence vector
}
\description{
Converts a co-located co-occurrence matrix (cocoma) to
a co-located co-occurrence vector (cocove)
}
\examples{
library(lopata)
library(raster)
data(raster_x, package = "comat")
data(raster_x_na, package = "comat")
raster_x = raster(raster_x)
raster_x_na = raster(raster_x_na)

plot(raster_x)
plot(raster_x_na)

coom = lop_cocoma(raster_x, raster_x_na)
coom

coov = lop_cocove(coom)
coov

}
