% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_coma.R
\name{lop_coma}
\alias{lop_coma}
\alias{lop_coma.RasterLayer}
\title{Create a co-occurrence matrix (coma)}
\usage{
lop_coma(x, neighbourhood, size, shift)

\method{lop_coma}{RasterLayer}(x, neighbourhood = 4, size = NULL,
  shift = NULL)
}
\arguments{
\item{x}{A RasterLayer with categories}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours:
4 (rook's case) or 8 (queen's case). The default is 4.}

\item{size}{Expressed in the numbers of cells, is a length of the side of a square-shaped block of cells. It defines the extent of a local pattern. If \code{size=NULL} calculations are performed for a whole area}

\item{shift}{Defines the shift between adjacent squares of cells along with the N-S and W-E directions. It describes the density (resolution) of the output grid. The resolution of the output map will be reduced to the original resolution multiplied by the shift. If shift=size the input map will be divided into a grid of non-overlapping square windows. Each square window defines the extent of a local pattern. If shift < size - results in the grid of overlapping square windows.}
}
\value{
A co-occurrence matrix
}
\description{
Create a co-occurrence matrix (coma)
}
\examples{
library(comat)
library(raster)
data(raster_x, package = "comat")
raster_x = raster(raster_x)
plot(raster_x)

com = lop_coma(raster_x)
com

com2 = lop_coma(raster_x, neighbourhood = 4, size = 2, shift = 2)
com2

}
