% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_cove.R
\name{lop_cove}
\alias{lop_cove}
\alias{lop_cove.coma}
\title{Create a co-occurrence vector (cove)}
\usage{
lop_cove(x, ordered, normalization)

\method{lop_cove}{coma}(x, ordered = TRUE, normalization = "none")
}
\arguments{
\item{x}{A matrix - an output of the \code{\link[=lop_coma]{lop_coma()}} function}

\item{ordered}{The type of pairs considered.
Either "ordered" (TRUE) or "unordered" (FALSE).
The default is TRUE.}

\item{normalization}{Should the output vector be normalized?
Either "none" or "pdf".
The "pdf" option normalizes a vector to sum to one.
The default is "none".}
}
\value{
A co-occurrence vector
}
\description{
Converts a co-occurrence matrix (coma) to
a co-occurrence vector (cove)
}
\examples{
library(comat)
library(raster)
data(raster_x, package = "comat")
raster_x = raster(raster_x)
plot(raster_x)

com = lop_coma(raster_x)
com

cov = lop_cove(com)
cov

}
