% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_incove.R
\name{lop_incove}
\alias{lop_incove}
\alias{lop_incove.incoma}
\title{Create an integrated co-occurrence vector (incove)}
\usage{
lop_incove(x, ordered, repeated, normalization)

\method{lop_incove}{incoma}(x, ordered = TRUE, repeated = TRUE,
  normalization = "none")
}
\arguments{
\item{x}{A matrix - an output of the \code{\link[=lop_incoma]{lop_incoma()}} function}

\item{ordered}{The type of pairs considered.
Either "ordered" (TRUE) or "unordered" (FALSE).
The default is TRUE.}

\item{repeated}{Should the repeated co-located co-occurrence matrices be used?
Either "ordered" (TRUE) or "unordered" (FALSE).
The default is TRUE.}

\item{normalization}{Should the output vector be normalized?
Either "none" or "pdf".
The "pdf" option normalizes a vector to sum to one.
The default is "none".}
}
\value{
An integrated co-occurrence vector
}
\description{
Converts an integrated co-occurrence matrix (incoma) to
an integrated co-occurrence vector (incove)
}
\examples{
library(comat)
library(raster)

l1 = raster(matrix(sample(1:2, size = 100, replace = TRUE), ncol = 10))
l2 = raster(matrix(sample(c(9, 6, 3), size = 100, replace = TRUE), ncol = 10))
x = stack(l1, l2, l1)

incom = lop_incoma(x)
incom

incov1 = lop_incove(incom)
incov1

incov2 = lop_incove(incom, ordered = FALSE)
incov2

}
