% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_thumbprint.R
\name{lop_thumbprint}
\alias{lop_thumbprint}
\title{Title}
\usage{
lop_thumbprint(
  ...,
  type,
  neighbourhood = 4,
  size = NULL,
  shift = NULL,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "none",
  fun = "mean",
  na_action = "replace"
)
}
\arguments{
\item{na_action}{}
}
\value{

}
\description{
Title
}
\examples{
library(comat)
library(raster)
landcover = raster(system.file("raster/landcover.tif", package = "lopata"))
plot(landcover)
landform = raster(system.file("raster/landform.tif", package = "lopata"))
plot(landform)
#npp = raster("npp.tif")
#plot(npp)
lop_thumbprint(landcover, type = "coma", threshold = 0.9)
lop_thumbprint(landcover, type = "cove", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocoma", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocove", threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecoma", threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecove", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incoma", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incove", threshold = 0.9)
lop_thumbprint(list(landcover, landform), type = "incove", threshold = 0.9)
}
