% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_to_dist.R
\name{lop_to_dist}
\alias{lop_to_dist}
\title{Calculate Distance Matrix on COMA}
\usage{
lop_to_dist(x, dist_fun, unit = "log2", p = NULL)
}
\arguments{
\item{x}{An object of class \code{cove}, \code{cocove}, \code{wecove}, or \code{incove}}

\item{dist_fun}{A distance/dissimilarity method used. All possible values can be found using
the \code{\link[philentropy:getDistMethods]{philentropy::getDistMethods()}} function}

\item{unit}{A character string specifying the logarithm unit that should be used to
compute distances that depend on log computations: \verb{"log", "log2", "log10"}.
The default is \code{"log"}.}

\item{p}{Power of the Minkowski distance}
}
\value{
An object of class `"dist"``
}
\description{
Calculates a distance matrix based on an object of class \code{cove}, \code{cocove}, \code{wecove}, or \code{incove}
}
\examples{
library(comat)
library(stars)
data(raster_x, package = "comat")
raster_x = st_as_stars(raster_x)
plot(raster_x)

cov = lop_thumbprint(raster_x, window_size = 2, type = "cove")

dist_cov = lop_to_dist(cov, dist_fun = "jensen-shannon")
dist_cov
}
