% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_wecove.R
\name{lop_wecove}
\alias{lop_wecove}
\alias{lop_wecove.wecoma}
\title{Create a weighted co-occurrence vector (wecove)}
\usage{
lop_wecove(x, ordered, normalization)

\method{lop_wecove}{wecoma}(x, ordered = TRUE, normalization = "none")
}
\arguments{
\item{x}{A matrix - an output of the \code{\link[=lop_wecoma]{lop_wecoma()}} function}

\item{ordered}{The type of pairs considered.
Either "ordered" (TRUE) or "unordered" (FALSE).
The default is TRUE.}

\item{normalization}{Should the output vector be normalized?
Either "none" or "pdf".
The "pdf" option normalizes a vector to sum to one.
The default is "none".}
}
\value{
A weighted co-occurrence vector
}
\description{
Converts a weighted co-occurrence matrix (wecoma) to
a weighted co-occurrence vector (wecove)
}
\examples{
library(comat)
library(raster)
data(raster_x, package = "comat")
data(raster_w_na, package = "comat")
raster_x = raster(raster_x)
raster_w_na = raster(raster_w_na)

plot(raster_x)
plot(raster_w_na)

wom = lop_wecoma(raster_x, raster_w_na)
wom

wov = lop_wecove(wom)
wov

}
