% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopata-package.R
\docType{package}
\name{lopata-package}
\alias{lopata}
\alias{lopata-package}
\title{lopata: Local Pattern Analysis}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nowosad/lopata}
  \item Report bugs at \url{https://github.com/Nowosad/lopata/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
