% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_add_quality.R
\name{lsp_add_quality}
\alias{lsp_add_quality}
\title{Calculates quality of clustering}
\usage{
lsp_add_quality(x, x_dist, regions = FALSE)
}
\arguments{
\item{x}{Object of class \code{lsp} - usually the output of the \code{lsp_add_clusters()} function}

\item{x_dist}{Object of class \code{dist} - usually the output of
the \code{lsp_to_dist()} function}

\item{regions}{Not implemented yet}
}
\description{
Calculates three quality metrics to evaluate spatial patterns' clustering:
(1) inhomogeneity - it measures a degree of mutual dissimilarity
between all objects in a cluster. This value is between 0 and 1,
where small value indicates that all objects in the cluster
represent consistent patterns so the cluster is pattern-homogeneous.
(2) isolation - it is an average distance between the focus cluster
and all of its neighbors. This value is between 0 and 1,
where large value indicates that the cluster
stands out from its surroundings.
(3) quality - overall quality of a cluster. It is calculated as
1 - (inhomogeneity / isolation). This value is also between 0 and 1,
where increased values indicate increased quality.
}
\examples{
# see examples of `lsp_add_clusters()`

}
\seealso{
lsp_add_clusters
}
