% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_add_spatial.R
\name{lsp_add_stars}
\alias{lsp_add_stars}
\alias{lsp_add_stars.default}
\alias{lsp_add_stars.lsp}
\title{Title}
\usage{
lsp_add_stars(x = NULL, window = NULL, window_size = NULL, window_shift = NULL)

\method{lsp_add_stars}{default}(x = NULL, window = NULL, window_size = NULL, window_shift = NULL)

\method{lsp_add_stars}{lsp}(x = NULL, window = NULL, window_size = NULL, window_shift = NULL)
}
\arguments{
\item{x}{}

\item{window_shift}{}
}
\description{
Title
}
\examples{
library(stars)
landform = read_stars(system.file("raster/landform.tif", package = "motif"))
plot(landform)
landform_lsp = lsp_add_stars(landform, window_size = 100)
plot(landform_lsp)

lc_cove = lsp_thumbprint(landform, type = "cove", window_size = 200, normalization = "pdf")
lc_cove_lsp = lsp_add_stars(lc_cove)
plot(lc_cove_lsp)
plot(lc_cove_lsp["na_prop"])

}
