% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_compare.R
\name{lsp_compare}
\alias{lsp_compare}
\title{Title}
\usage{
lsp_compare(
  x,
  y,
  type,
  dist_fun,
  window = NULL,
  window_size = NULL,
  window_shift = NULL,
  neighbourhood = 4,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "pdf",
  wecoma_fun = "mean",
  wecoma_na_action = "replace"
)
}
\arguments{
\item{wecoma_na_action}{}
}
\value{

}
\description{
Title
}
\examples{
library(stars)

lc15 = read_stars(system.file("raster/landcover2015.tif", package = "motif"))
lc01 = read_stars(system.file("raster/landcover2001.tif", package = "motif"))
lf = read_stars(system.file("raster/landform.tif", package = "motif"))
ecoregions = read_stars(system.file("raster/ecoregions.tif", package = "motif"))

s1 = lsp_compare(lc01, lc15, type = "cove", dist_fun = jensen_shannon, threshold = 0.9)
s1b = lsp_compare(lc01, lc15, type = "cove", dist_fun = jensen_shannon, window = ecoregions, threshold = 0.9)
s1c = lsp_compare(lc01, lc15, type = "cove", dist_fun = jensen_shannon, window_size = 1000, threshold = 0.9)
s2 = lsp_compare(c(lc01, lf), c(lc15, lf), type = "cocove", dist_fun = jensen_shannon, threshold = 0.9)
s3 = lsp_compare(c(lc01, lf), c(lc15, lf), type = "wecove", dist_fun = jensen_shannon, threshold = 0.9)
s4 = lsp_compare(c(lc01, lf), c(lc15, lf), type = "incove", dist_fun = jensen_shannon, threshold = 0.9)

library(sf)
eco_sf = read_sf(system.file("vector/ecoregions.gpkg", package = "motif"))
eco_sf = dplyr::bind_cols(eco_sf, s1b)
}
