% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_extract.R
\name{lsp_extract}
\alias{lsp_extract}
\title{Extracts a local landscape}
\usage{
lsp_extract(x, window, id)
}
\arguments{
\item{x}{Object of class \code{stars} or \code{stars_proxy}.}

\item{window}{Specifies areas for analysis. It can be either: \code{NULL}, a numeric value, or an \code{sf} object. If \code{window=NULL} calculations are performed for a whole area. If the \code{window} argument is numeric, it is a length of the side of a square-shaped block of cells. Expressed in the numbers of cells, it defines the extent of a local pattern. If an \code{sf} object is provided, each feature (row) defines the extent of a local pattern. The \code{sf} object should have one attribute (otherwise, the first attribute is used as an id).}

\item{id}{Id of the local landscape - it is possible to find in the output of \code{lsp_thumbprint()}, \code{lsp_search()}, \code{lsp_compare()}, or \code{lsp_add_clusters()}.}
}
\description{
Extracts a local landscape from categorical raster data based on its id and provided \code{window} argument.
}
\examples{
library(stars)
landform = read_stars(system.file("raster/landform.tif", package = "motif"))
ecoregions = read_sf(system.file("vector/ecoregions.gpkg", package = "motif"))
plot(lsp_extract(x = landform, window = 100, id = 1895))
plot(lsp_extract(x = landform, window = ecoregions["id"], id = 7))
}
