% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_search.R
\name{lsp_search}
\alias{lsp_search}
\alias{lsp_search.stars}
\title{Title}
\usage{
lsp_search(
  x,
  y,
  type,
  dist_fun,
  window = NULL,
  window_size = NULL,
  window_shift = NULL,
  neighbourhood = 4,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "pdf",
  wecoma_fun = "mean",
  wecoma_na_action = "replace",
  ...
)

\method{lsp_search}{stars}(
  x,
  y,
  type,
  dist_fun,
  window = NULL,
  window_size = NULL,
  window_shift = NULL,
  neighbourhood = 4,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "pdf",
  wecoma_fun = "mean",
  wecoma_na_action = "replace",
  ...
)
}
\arguments{
\item{dist_fun}{}
}
\value{

}
\description{
Title
}
\examples{
library(stars)

lc = read_stars(system.file("raster/landcover2015.tif", package = "motif"))
#lf = read_stars(system.file("raster/landform.tif", package = "motif"))

ext = st_bbox(c(xmin = -249797.344531127, xmax = -211162.693944285,
                ymin = -597280.143035389, ymax = -558645.492448547),
                crs = st_crs(lc))

#ecoregions = read_stars(system.file("raster/ecoregions.tif", package = "motif"))
#plot(ecoregions)

lc_ext = lc[ext]
#lf_ext = lf[ext]

s1 = lsp_search(lc_ext, lc, type = "cove",
  dist_fun = "jensen-shannon", threshold = 0.9)
#s2 = lsp_search(c(lc_ext, lf_ext), c(lc, lf), type = "cocove",
#  dist_fun = "jensen-shannon", threshold = 0.9)
#s3 = lsp_search(c(lc_ext, lf_ext), c(lc, lf), type = "wecove",
#  dist_fun = "jensen-shannon", threshold = 0.9)
#s4 = lsp_search(c(lc_ext, lf_ext), c(lc, lf), type = "incove",
#  dist_fun = "jensen-shannon", threshold = 0.9)
#s5 = lsp_search(lf_ext, lf, type = "cove",
#  dist_fun = "jensen-shannon", threshold = 0.5, window_size = 250)
#s6 = lsp_search(lc_ext, lc, type = "cove",
#  dist_fun = "jensen-shannon", threshold = 0.5, window = ecoregions)
}
