% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_window.R
\name{prepare_window}
\alias{prepare_window}
\title{Prepares window* arguments (internal function)}
\usage{
prepare_window(x, window)
}
\arguments{
\item{x}{Object of class \code{stars} or \code{stars_proxy}}

\item{window}{Specifies areas for analysis. It can be either: \code{NULL}, a numeric value, or an \code{sf} object. If \code{window=NULL} calculations are performed for a whole area. If the \code{window} argument is numeric, it is a length of the side of a square-shaped block of cells. Expressed in the numbers of cells, it defines the extent of a local pattern. If an \code{sf} object is provided, each feature (row) defines the extent of a local pattern. The \code{sf} object should have one attribute (otherwise, the first attribute is used as an id).}
}
\value{
A list with \code{window}, \code{window_size}, and \code{window_shift}
}
\description{
Prepares window* arguments (internal function)
}
