% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dist.R
\name{to_dist}
\alias{to_dist}
\alias{to_dist.cove}
\alias{to_dist.cocove}
\alias{to_dist.wecove}
\alias{to_dist.incoma}
\title{Title}
\usage{
to_dist(x, method, unit, p)

\method{to_dist}{cove}(x, method, unit = "log", p = NULL)

\method{to_dist}{cocove}(x, method, unit = "log", p = NULL)

\method{to_dist}{wecove}(x, method, unit = "log", p = NULL)

\method{to_dist}{incoma}(x, method, unit = "log", p = NULL)
}
\arguments{
\item{method}{\code{\link[philentropy:getDistMethods]{philentropy::getDistMethods()}}}

\item{p}{}
}
\description{
Title
}
\examples{

}
