% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_classes.R
\name{determine_classes}
\alias{determine_classes}
\title{Determine unique classes (internal function)}
\usage{
determine_classes(x, window)
}
\arguments{
\item{x}{\itemize{
\item a \code{stars} or \code{stars_proxy} object
}}

\item{window}{\itemize{
\item a windows argument from \code{lsp_signature()}, \code{lsp_search()}, or \code{lsp_compare()}
}}
}
\value{
A list with vector of numbers (unique classes)
}
\description{
Determine unique classes (internal function)
}
