% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cove.R
\name{get_cove}
\alias{get_cove}
\title{Create a co-occurrence vector (wecove)}
\usage{
get_cove(x, ordered = TRUE)
}
\arguments{
\item{x}{A matrix - an output of the \code{\link[=get_coma]{get_coma()}} function}

\item{ordered}{The type of pairs considered.
Either ordered (TRUE) or unordered (FALSE).
The default is TRUE.}
}
\value{
A co-occurrence vector
}
\description{
Converts a co-occurrence matrix (wecoma) to
a co-occurrence vector (wecove)
}
\examples{
library(comat)
library(raster)
data(x, package = "comat")

com = get_coma2(x)
com

cov = get_cove(com$matrix[[1]])
cov

get_cove2(com)
}
