% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_compare.R
\name{lop_compare}
\alias{lop_compare}
\title{Title}
\usage{
lop_compare(
  x,
  y,
  type,
  dist_fun,
  window = NULL,
  window_size = NULL,
  window_shift = NULL,
  neighbourhood = 4,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "pdf",
  wecoma_fun = "mean",
  wecoma_na_action = "replace"
)
}
\arguments{
\item{wecoma_na_action}{}
}
\value{

}
\description{
Title
}
\examples{
library(raster)
library(philentropy)
ext = extent(c(xmin = -249797.344531127, xmax = -211162.693944285,
               ymin = -597280.143035389, ymax = -558645.492448547))

lc = raster(system.file("raster/landcover.tif", package = "lopata"))
lf = raster(system.file("raster/landform.tif", package = "lopata"))

s1 = lop_compare(list(lc), list(lc), type = "cove", dist_fun = jensen_shannon, threshold = 0.9)
s2 = lop_compare(list(lc, lf), list(lc, lf), type = "cocove", dist_fun = jensen_shannon, threshold = 0.9)
s3 = lop_compare(list(lc, lf), list(lc, lf), type = "wecove", dist_fun = jensen_shannon, threshold = 0.9)
s4 = lop_compare(list(lc, lf), list(lc, lf), type = "incove", dist_fun = jensen_shannon, threshold = 0.9)
}
