% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_search.R
\name{lop_search}
\alias{lop_search}
\title{Title}
\usage{
lop_search(
  x,
  y,
  type,
  dist_fun,
  window,
  window_size = NULL,
  window_shift = NULL,
  neighbourhood = 4,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "pdf",
  wecoma_fun = "mean",
  wecoma_na_action = "replace"
)
}
\arguments{
\item{dist_fun}{}
}
\value{

}
\description{
Title
}
\examples{
library(stars)
library(philentropy)

lc = read_stars(system.file("raster/landcover2015.tif", package = "lopata"))
lf = read_stars(system.file("raster/landform.tif", package = "lopata"))

ext = st_bbox(c(xmin = -249797.344531127, xmax = -211162.693944285,
                ymin = -597280.143035389, ymax = -558645.492448547),
                crs = st_crs(lc))


lc_ext = lc[ext]
lf_ext = lf[ext]

s1 = lop_search(lc_ext, lc, type = "cove", dist_fun = jensen_shannon, threshold = 0.9)
s2 = lop_search(c(lc_ext, lf_ext), c(lc, lf), type = "cocove", dist_fun = jensen_shannon, threshold = 0.9)
s3 = lop_search(c(lc_ext, lf_ext), c(lc, lf), type = "wecove", dist_fun = jensen_shannon, threshold = 0.9)
s4 = lop_search(c(lc_ext, lf_ext), c(lc, lf), type = "incove", dist_fun = jensen_shannon, threshold = 0.9)
}
