% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_thumbprint.R
\name{lop_thumbprint}
\alias{lop_thumbprint}
\title{Title}
\usage{
lop_thumbprint(
  ...,
  type,
  window,
  window_size = NULL,
  window_shift = NULL,
  neighbourhood = 4,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "none",
  wecoma_fun = "mean",
  wecoma_na_action = "replace"
)
}
\arguments{
\item{na_action}{}
}
\value{

}
\description{
Title
}
\examples{
library(comat)
library(raster)
landcover = raster(system.file("raster/landcover.tif", package = "lopata"))
plot(landcover)
landform = raster(system.file("raster/landform.tif", package = "lopata"))
plot(landform)
#npp = raster("npp.tif")
#plot(npp)
ecoregions = raster(system.file("raster/ecoregions.tif", package = "lopata"))
plot(ecoregions)
lop_thumbprint(landcover, type = "coma", threshold = 0.9)
lop_thumbprint(landcover, type = "cove", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocoma", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocove", threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecoma", threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecove", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incoma", threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incove", threshold = 0.9)
lop_thumbprint(list(landcover, landform), type = "incove", threshold = 0.9)

lop_thumbprint(landcover, type = "coma", window_size = 100, window_shift = 100, threshold = 0.9)
lop_thumbprint(landcover, type = "cove", window_size = 100, window_shift = 100, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocoma", window_size = 100, window_shift = 100, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocove", window_size = 100, window_shift = 100, threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecoma", window_size = 100, window_shift = 100, threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecove", window_size = 100, window_shift = 100, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incoma", window_size = 100, window_shift = 100, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incove", window_size = 100, window_shift = 100, threshold = 0.9)
lop_thumbprint(list(landcover, landform), type = "incove", window_size = 100, window_shift = 100, threshold = 0.9)

lop_thumbprint(landcover, type = "coma", window = ecoregions, threshold = 0.9)
lop_thumbprint(landcover, type = "cove", window = ecoregions, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocoma", window = ecoregions, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "cocove", window = ecoregions, threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecoma", window = ecoregions, threshold = 0.9)
#lop_thumbprint(landcover, npp, type = "wecove", window = ecoregions, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incoma", window = ecoregions, threshold = 0.9)
lop_thumbprint(landcover, landform, type = "incove", window = ecoregions, threshold = 0.9)
lop_thumbprint(list(landcover, landform), type = "incove", window = ecoregions, threshold = 0.9)
}
