% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lop_to_dist.R
\name{lop_to_dist}
\alias{lop_to_dist}
\alias{lop_to_dist.cove}
\alias{lop_to_dist.cocove}
\alias{lop_to_dist.wecove}
\alias{lop_to_dist.incoma}
\title{Calculate Distance Matrix on COMA}
\usage{
lop_to_dist(x, method, unit, p)

\method{lop_to_dist}{cove}(x, method, unit = "log", p = NULL)

\method{lop_to_dist}{cocove}(x, method, unit = "log", p = NULL)

\method{lop_to_dist}{wecove}(x, method, unit = "log", p = NULL)

\method{lop_to_dist}{incoma}(x, method, unit = "log", p = NULL)
}
\arguments{
\item{x}{An object of class \code{cove}, \code{cocove}, \code{wecove}, or \code{incove}}

\item{method}{A distance/dissimilarity method used. All possible values can be found using
the \code{\link[philentropy:getDistMethods]{philentropy::getDistMethods()}} function}

\item{unit}{A character string specifying the logarithm unit that should be used to
compute distances that depend on log computations: \code{"log", "log2", "log10"}.
The default is \code{"log"}.}

\item{p}{Power of the Minkowski distance}
}
\value{
An object of class `"dist"``
}
\description{
Calculates a distance matrix based on an object of class \code{cove}, \code{cocove}, \code{wecove}, or \code{incove}
}
\examples{
library(comat)
library(raster)
data(raster_x, package = "comat")
raster_x = raster(raster_x)
plot(raster_x)

com = lop_coma(raster_x, size = 2)
com

cov = lop_cove(com)
cov

dist_cov = lop_to_dist(cov, method = "jensen-shannon")
dist_cov
}
