% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_add_clusters.R
\name{lsp_add_clusters}
\alias{lsp_add_clusters}
\title{Adds clusters' ids to a lsp object}
\usage{
lsp_add_clusters(x, clust, output = "sf", window = NULL)
}
\arguments{
\item{x}{Object of class \code{lsp} - usually the output of
the \code{lsp_thumbprint()} function}

\item{clust}{Vector containing an id value for each row in \code{x}}

\item{output}{The class of the output. Either \code{"stars"} or \code{"sf"}}

\item{window}{Specifies areas for analysis. It can be either: \code{NULL} or an \code{sf} object. If \code{window=NULL} calculations are performed based on the metadata from \code{x}. If an \code{sf} object is provided, each feature (row) defines the extent of a local pattern. The \code{sf} object should have one attribute (otherwise, the first attribute is used as an id).}
}
\description{
Adds clusters' ids to a lsp object.
The output can be of \code{stars} or \code{sf} class.
See examples.
}
\examples{
library(stars)
landform = read_stars(system.file("raster/landform.tif", package = "motif"))
landform_cove = lsp_thumbprint(landform,
                               type = "cove",
                               window = 200,
                               normalization = "pdf")

landform_dist = lsp_to_dist(landform_cove,
                            dist_fun = "jensen-shannon")

landform_hclust = hclust(landform_dist, method = "ward.D2")
plot(landform_hclust)

clusters = cutree(landform_hclust, k = 6)

safe_pal = c("#88CCEE", "#CC6677", "#DDCC77",
             "#117733", "#332288", "#888888")

landform_grid_stars = lsp_add_clusters(landform_cove, clusters)
plot(landform_grid_stars["clust"], col = safe_pal)

landform_grid_starsq = lsp_add_quality(landform_grid_stars,
                                       landform_dist)
plot(landform_grid_starsq["quality"])
}
