% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_add_quality.R
\name{lsp_add_quality}
\alias{lsp_add_quality}
\title{Title}
\usage{
lsp_add_quality(x, x_dist, clust_var, regions = FALSE)
}
\arguments{
\item{x}{}

\item{regions}{}
}
\description{
Title
}
\examples{
library(stars)
library(rcartocolor)

landcover = read_stars(system.file("raster/landcover2015.tif", package = "motif"))
# plot(landcover)
lc_cove = lsp_thumbprint(landcover, type = "cove", window_size = 1000, normalization = "pdf")
lc_dist = lsp_to_dist(lc_cove, dist_fun = "jensen-shannon")
lc_hclust = hclust(lc_dist, method = "ward.D2")
clusters = cutree(lc_hclust, k = 12)

lc_grid = lsp_add_clusters(lc_cove, clusters)
plot(lc_grid["clust"], col = carto_pal(12, "Safe"))

lc_grid = lsp_add_quality(lc_grid, lc_dist, "clust")
plot(lc_grid["inhomogeneity"])
plot(lc_grid["isolation"])
plot(lc_grid["quality"])

}
