% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_add_spatial.R
\name{lsp_add_sf}
\alias{lsp_add_sf}
\alias{lsp_add_sf.default}
\alias{lsp_add_sf.lsp}
\title{Title}
\usage{
lsp_add_sf(x = NULL, window = NULL, window_size = NULL, window_shift = NULL)

\method{lsp_add_sf}{default}(x = NULL, window = NULL, window_size = NULL, window_shift = NULL)

\method{lsp_add_sf}{lsp}(x = NULL, window = NULL, window_size = NULL, window_shift = NULL)
}
\arguments{
\item{x}{}

\item{window_shift}{}
}
\description{
Title
}
