% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_thumbprint.R
\name{lsp_thumbprint}
\alias{lsp_thumbprint}
\title{Title}
\usage{
lsp_thumbprint(
  x,
  type,
  window = NULL,
  window_size = NULL,
  window_shift = NULL,
  neighbourhood = 4,
  threshold = 0.5,
  ordered = TRUE,
  repeated = TRUE,
  normalization = "none",
  wecoma_fun = "mean",
  wecoma_na_action = "replace",
  classes = NULL
)
}
\arguments{
\item{na_action}{}
}
\value{

}
\description{
Title
}
\examples{
library(stars)

landcover = read_stars(system.file("raster/landcover2015.tif", package = "motif"))
#plot(landcover)
#landform = read_stars(system.file("raster/landform.tif", package = "motif"))
#plot(landform)
#ecoregions = read_stars(system.file("raster/ecoregions.tif", package = "motif"))
#plot(ecoregions)

lsp_thumbprint(landcover, type = "coma", threshold = 0.9)
#lsp_thumbprint(landcover, type = "cove", threshold = 0.9)
#lsp_thumbprint(landcover, type = "cove", threshold = 0.9, classes = 10)
#lsp_thumbprint(c(landcover, landform), type = "incove", threshold = 0.9)

#lsp_thumbprint(landcover, type = "coma", window_size = 100, window_shift = 100, threshold = 0.9)

#lsp_thumbprint(landcover, type = "coma", window = ecoregions, threshold = 0.9)
}
